Refactoring: import 경로 변경 및 Comment Entity간 cascade 관계 수정
